Refactor <TopPanel /> as functional component
