+++ resources/assets/js/components/dashboard/SidePanel.js
